Build project with Go 1.20 and Go 1.21
As part of the upgrade to Go 1.21, the Go toolchain now requires the
`go` directive to match the maximum Go version in use in dependencies.

This leads to any transitive dependency on this library to result in a
requirement of the consuming project moving to Go 1.21.

Although in a lot of cases this may not be as problematic, it forces
consumers to migrate, which we don't need to do in this case.

By making sure we build across both the current venison and the next
version of Go - taking care to only support the two major Go versions in
use by the Go team - we can hopefully pick up on similar issues.

This is a slightly more complex setup than just specifying the version
manually, but to avoid the risk of not updating it in all the places, we
can at a minimum use the `go.mod` version for the current targeted
version.

We can make sure to not `fail-fast` so we can see each of the builds'
results, instead of only seeing the first to fail.